requirements.txt
streamlit
google-generativeai
numpy
scikit-learn
sentence-transformers
python-dotenv
faiss-cpu  
